undoing capacitance and material property smoothing (reverting back to commit: 228211f) for C code
Testing demonstrates that the C code and Py code both truncate values in lookup tables,
so it doesn’t seem that this new smoothing approach should cause problems or differences
in that regard.  But the performance is greatly reduced for the C code (orders of magnitude),
which might be because of the increased number of lookups.  Evidently we are more limited
by lookups than evaluating if statements.