Monitor: placeholder PETSc binary output and python reader
Dump three vectors into a PETSc binary file in output/ at each
monitoring step.

The first vector can be used to hold relevant data - here the timestep
and dtmacro.

The second vector is the solution, and the third is phi_s (different
number of points).

The files are named with the timestep (naming with the number of years
caused overwriting in my tests with very short dtmacro).

I didn't see an obvious way to extract the vector names, so
unfortunately this relies on the python script knowing the order that
the C code dumped things.

The python script just has a hard-coded set of steps to import, and
prints out the data it finds. There isn't much error checking!