Lookup tables: add -curves command line option
These changes are to allow us to avoid recompiling the code just to change the data table lookups.

Logic is added to lookup.c which looks for -curves stixrude2009, -curves andrault2011 (or -curves default) on the command line, and then loads the appropriate files.

This should be easily extended to define other sets of files to load.