Interp1d : fix memory error
The previous logic allowed for an index of -1 to be computed. This was
actually happening in the case of looking up a pressure value. both x
and xmin were equal to zero, so xa[-1] was being accessed.

The change updates the behavior when a value exactly matches one in the
interpolation table, by changing > to >= .

A parallel change is made in searching the y table.

In both cases, if performance ever becomes an issue, this algorithm
could be replaced with bisection. However, these tables are so small
that it's not obvious that this naive algorithm would even be the less
efficient choice.