Ctx: Populate a DMComposite object to manage the solution vector
We want to be able to solve with more-arbitrary fields in the future, so
we define multiple domains to hold the different parts of the solution.

Note that the solver doesn't know anything about DimensionalisableField,
but DimensionalisableField is designed to play nicely with DMComposite.

This does not remove the "aug" logic yet, though it does use the new DMComposite to create the solution vector in main.c (brittle, since this depends on the exact structure of the fields!)