Add PostStep and RollBack functionality (with placeholders)
Use -activate_rollback to enable TSRollBack() for any TS implementation,
in particular TSSUNDIALS. This is achieved by using the technique as in
this example: https://bitbucket.org/psanan/ts_generic_rollback_example

Use -activate_poststep to add a post-step hook, allowing arbitary code
(in the PostStep() function) to be executed after each time step.

Add placeholder implementation for a post-step hook which compares
data stored in a Ctx and rolls back, terminates TSSolve(), and ends
macro stepping.

This gives us a the beginnings of an event detector. Indeed, more logic
could be added to not simply terminate the solution process, but to
restart with a smaller timestep.

Test with e.g.

   valgrind ./spider -activate_poststep -activate_rollback -dtmacro_years 0.001 -ts_monitor