Volatiles: generalize to arbitrary number/names
Replace the hard-coded CO2 and H2O volatile names with a 
mechanism to read a list of names from the command line,
like

   -volatile_names CO2,H2O

This sets a "prefix" for each volatile, which is used to process further options.

There is still a maximum number of volatiles allowed, defined
by SPIDER_MAX_VOLATILE_SPECIES, but this macro should be avoided whenever possible. It is left for now to avoid introducing more construction/destruction boilerplate.

Update mechanism to set these options to throw errors if a parameter without a default isn't set with a command-line option.

Change many loops to iterate over the now-variable number of volatiles (in AtmosphereParameters.n_volatiles).

Update examples to use existing CO2+H2O setup.

Add TODOS for

* better default values for volatiles
* output of volatile parameters with other parameters