Volatile reactions: refactor and generalize
Replace the machinery for the treatment of volatile reactions in the
liquid mantle with a more generalizable version of the same approach,
in particularly allowing for more than a single reaction.

This is currently restricted to the following "simple" case:
- only two volatiles per reaction
- an equilibrium condition specified by two numbers epsilon0 and
epsilon1, such that

    epsilon0 * volatile0.p + epsilon1 * volatile1.p = 0

This implementation relies on command line options. See
examples/kutra_2019/constML_reactions/with_reactions/constML_H2O_H2_new.opt