EOSRTpress: remove questions in comments
PetscCbrtReal is newer than our version of PETSc.

If an expression makes sense with complex scalars, then use the "Scalar"
functions. If it only makes sense with reals, use the "Real" ones.
In practice, just use the "Scalar" ones unless you have a reason not to
(and it won't ever come up with this code, as we won't use complex
scalars, so Real and Scalar will always be the same).